setup eslint, refactor, etc
