+++ .eslintrc.json
@@ -1,0 +1,25 @@
+{
+    "env": {
+        "browser": true,
+        "es2021": true
+    },
+    "extends": [
+        "airbnb-base",
+        "prettier"
+    ],
+    "plugins": ["prettier"],
+    "parserOptions": {
+        "ecmaVersion": "latest",
+        "sourceType": "module"
+    },
+    "rules": {
+        "prettier/prettier": ["error", { "singleQuote": true }],
+        "no-continue": "off",
+        "no-use-before-define": [
+            "error",
+            { "functions": false }
+        ],
+        "import/extensions": "off"
+  
+    }
+}
