+++ src/assets/boo.png
