+++ src/index.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
   <head>
-    <title>Daily Chat + Emoji recording with VCS</title>
+    <title>Daily chat + reaction recording with VCS</title>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link href="./style.css" rel="stylesheet" />  
@@ -12,7 +12,7 @@
         your browser.
     </noscript>
     <div id="header">
-      <span>custom session data sample</span><span>powered by <a href="https://docs.daily.co">//daily</a></span>
+      <span>Daily chat + reaction recording with VCS</span><span>powered by <a href="https://docs.daily.co">//daily</a></span>
     </div>
     <div id="container">
         <div id="controls">
